[*] move `ListenAndServe()` and interceptors to server.go
